docs: update release process
